pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advdef-train"
version = "0.0.1"
description = "Transfer-learning experiments that integrate adv-it-defenses"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "torch>=2.1",
    "torchvision>=0.16",
    "tqdm>=4.66",
    "Pillow>=10.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[tool.setuptools]
packages = ["training"]
